Rename: temporarily rename common folder name
